Add x64 and Direct3D 11 client support

+++ src/engine/shared/library/sharedFoundation/src/shared/AutoDeltaNetworkIdPackedMap.h
@@ -10,6 +10,7 @@
 
 // ======================================================================
 
+#include "NetworkIdArchive.h"
 #include "Archive/AutoDeltaMap.h"
 #include <stdio.h>
 #include "../../../../../../engine/shared/library/sharedFoundation/include/public/sharedFoundation/NetworkId.h"
@@ -32,9 +33,9 @@ namespace Archive
 		char temp[200];
 		AutoDeltaMap<NetworkId, int>::Command c;
 
-		Archive::put(target, countCharacter(buffer,':'));
-		Archive::put(target, static_cast<size_t>(0)); // baselineCommandCount
-		
+		Archive::put(target, static_cast<int32_t>(countCharacter(buffer, ':')));
+		Archive::put(target, static_cast<int32_t>(0)); // baselineCommandCount
+
 		int tempPos = 0;
 		for (std::string::const_iterator i=buffer.begin(); i!=buffer.end(); ++i)
 		{
@@ -63,8 +64,8 @@ namespace Archive
 		char temp[200];
 		
 		AutoDeltaMap<NetworkId, int>::Command c;
-		size_t commandCount;
-		size_t baselineCommandCount;
+		int32_t commandCount;
+		int32_t baselineCommandCount;
 
 		Archive::get(source, commandCount);
 		Archive::get(source, baselineCommandCount);
@@ -75,7 +76,7 @@ namespace Archive
 		}
 		else
 		{
-			for (size_t i = 0; i < commandCount; ++i)
+			for (int32_t i = 0; i < commandCount; ++i)
 			{
 				Archive::get(source, c.cmd);
 				Archive::get(source, c.key);
